src/technical_Admin/WorklistAssignerListTest.java
package technical_Admin;

import org.testng.Assert;
import org.testng.ITestResult;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeSuite;
import org.testng.annotations.Test;

import com.relevantcodes.extentreports.ExtentReports;
import com.relevantcodes.extentreports.ExtentTest;
import com.relevantcodes.extentreports.LogStatus;


import TestCases.WebDriverPass;
import quality_Admin.FormTypeLibrary;
import technical_Admin.WorklistAssignerList;

public class WorklistAssignerListTest extends WebDriverPass{
	protected ExtentReports report;
	protected ExtentTest test;
	
	WorklistAssignerList w=new WorklistAssignerList();
 
  @Test(priority=33)
  public void WorklistAssignerListindex() throws Exception  {
	
	  test=report.startTest("WorklistAssignerListindex");	   
 w.WorklistAssignerListindex(driver);
 Assert.assertTrue(true);
 test.log(LogStatus.PASS, "WorklistAssignerListindex");         
  }
  

  @Test(priority=34)
  public void Alert() throws Exception  {
	  test=report.startTest("Alert");	 
 w.Alert(driver);
 Assert.assertTrue(true);
 test.log(LogStatus.PASS, "Alert");    
}

  @AfterMethod
	public void afterMethod(ITestResult result) {
	    if (result.getStatus() == ITestResult.FAILURE) {
	        test.log(LogStatus.FAIL, "Test failed " + result.getThrowable());
	    } else if (result.getStatus() == ITestResult.SKIP) {
	        test.log(LogStatus.SKIP, "Test skipped " + result.getThrowable());
	    } else {
	        test.log(LogStatus.PASS, "Test passed");
	    }
	    report.endTest(test);
	    report.flush();
}
@BeforeSuite
public void beforeSu() {
	report = FormTypeLibrary.getReporter("C:\\Users\\Priti\\workspace\\JiBeAutomation\\Report\\TechnicalAdmin.html");		  }

/* @AfterSuite
public void afterSu() {
    report.close();
}*/
  
}